Поправил MuiContainer паддинги(левый, правый) и breakpoint на 1440px, чтобы было по умолчанию. Переделал логику фильтрации WorkCategories на главной странице
