+++ src/components/DesinersCarousel/DesinersCarousel.scss
@@ -1,8 +1,6 @@
 .desinersCarousel {
-  max-width: 1440px !important;
-
   &__list {
-    margin: 0;
+    margin: 0 auto;
     padding: 0;
     list-style-type: none;
     gap: 24px;
