+++ src/components/DesinersCategories/DesinersCategories.scss
@@ -1,8 +1,4 @@
 .desinersCategories {
-  max-width: 1440px !important;
-
-  padding: 24px 0;
-
   &__list {
     margin: 0;
     padding: 0;
@@ -16,7 +12,7 @@
 
     display: flex;
     flex-direction: column;
-    gap: 20px;
+    gap: 19px;
 
     text-align: center;
 
@@ -30,6 +26,8 @@
       border-radius: 12px;
       transform-origin: 0% 0%;
       transition: transform 0.4s linear;
+      box-shadow: 0px 4px 16px 0px rgba(51, 51, 51, 0.08),
+        0px 4px 4px 0px rgba(51, 51, 51, 0.04);
     }
 
     &__content {
@@ -62,6 +60,8 @@
 
     &:hover &__image {
       transform: scale(1.03);
+      box-shadow: 0px 4px 24px 0px rgba(51, 51, 51, 0.24),
+        0px 4px 4px 0px rgba(51, 51, 51, 0.04);
     }
   }
