+++ src/components/Feed/Feed.scss
@@ -1,7 +1,5 @@
 .feed {
-  max-width: 1440px !important;
-
-  padding-top: 60px;
+  padding-top: 82px;
   padding-bottom: 100px;
 
   &__list {
