+++ src/components/WorkCategories/WorkCategories.scss
@@ -1,6 +1,4 @@
 .workCategories {
-  max-width: 1440px !important;
-
   padding: 0;
   margin: 0 0 32px 0;
   list-style-type: none;
