+++ src/pages/MainPage/MainPage.scss
@@ -16,6 +16,6 @@
   &__container {
     min-height: 100vh;
     font-family: "Roboto", "Helvetica", "Arial", sans-serif;
-    gap: 76px;
+    gap: 53px;
   }
 }
